fix: Metric 엔티티에 user 필드 추가

+++ src/main/java/com/example/llmn/controller/DTO/MetricResponse.java
@@ -32,10 +32,10 @@ public class MetricResponse {
 
     public record FindCurrentMetricDTO(
             double cpuUsage,
-            long totalMemory,
-            long usedMemory,
-            long networkReceived,
-            long networkSent){}
+            double totalMemory,
+            double usedMemory,
+            double networkReceived,
+            double networkSent){}
 
     public record CommandDTO(
             String privateKeyPath,
